feat(weather): display temperature as an integer in handler

+++ src/handlers/weather.rs
@@ -167,7 +167,7 @@ fn format_weather_message(city: &str, response: &Forecast) -> String {
         {emoji} {description}\n\n\
         <i>Гарного дня!</i> ☀️",
         city=city,
-        temp=response.main.temp,
+        temp=response.main.temp as i32,
         emoji=emoji,
         description=description,
     )
